generalize sensorController
a bit of renaming

more refactors

more refactor

rename tests

clean up includes

rename

move stuff around

more refactors

more cleanup

remove unused var
