--- test/stubs/mock_aanderaaController.cpp
@@ -1,3 +1,0 @@
-#include "aanderaaController.h"
-
-TaskHandle_t aanderaa_controller_task_handle = NULL;
