+++ test/stubs/mock_sensorController.cpp
@@ -1,0 +1,3 @@
+#include "sensorController.h"
+
+TaskHandle_t sensor_controller_task_handle = NULL;
